fix: unset env vars PS1 and PS2 that could interfere with shell runner
The PS1 and PS2 are the environment variables that define the prompts
for the shell runner.

As soon as possible, the runner will change those but the shell spawned
still will use them during the first stages before the runner have a
chance to change them.

During this period, the shell will use the PS1 and PS2 inherited from
the parent environment. It is common to have those customized running
arbitrary commands (pretty prompts) which will end up being executed
in the context of the runner.

Unsetting these two should prevent that.

Note: no necessary PS1 and PS2 are the only variables inherited that may
have side effects but certainly the two most common.