fix: prevent RC on Options doing a deep copy
The Config object relays on Options.copy to ensure that the multiple
Config cloned objects remain independent each other and thread safe.

Because Options.copy was not deep, it could be happen that a mutable
value of Options may be changed by one thread and visible by another.

To prevent this we do a deep copy.