test: fix quote of ! for shell example that may not take literal
The '!' is a special symbol in Bash. In some versions if the '!' is not
followed by something (like in 'Boom!'), it is taken literal (as it
is expected in the examples).

In some versions of Bash however, it may print an error. To remove the
possibility, we quote this properly.