feat: pass shell's pid to send it a SIGWINCH
The session id is the pid of the leader process of the current session.
It may be the shell handling the autocomplete in foreground but not necessary.

A leader shell may spawn another shell and it is this shell who owns the
terminal and it is handling the autocomplete. Sending a SIGWINCH to the
former shell makes no effect.

The only reliable way to know the pid of the correct shell is
retrieving it from the autocomplete script and passing it to byexample
via an environment variable.