fix: increase geometry of C++ (API change)
This change may break some C++ examples if their output is larger
than 128 cols or rows.

The limit is extended to 2048 cols and 1024 rows using the optimized
termscraper's Screen.

The runner also delegates to the parent class the output processing
forcing the echo filtering with _is_echo_filtering_enforced() in True