fix: unset PAGER, LESS, PROMPT_COMMAND env vars in Shell
This shell variable may interfere with the examples as they are executed
quite often and they are probably tweaked by the user for his/her
shell and not for byexample's shell.

Because we are not loading any rc or profile file, it is unlikely that
these variables are set in a byexample shell session however it is
better to be sure.